Add sync e2e test
