+++ endtoend/beacon_node.go
@@ -1,34 +1,20 @@
 package endtoend
 
 import (
-	"bufio"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
 	"path"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/bazelbuild/rules_go/go/tools/bazel"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/pkg/errors"
 	ev "github.com/prysmaticlabs/prysm/endtoend/evaluators"
 	"github.com/prysmaticlabs/prysm/shared/params"
 )
 
-// BeaconNodeInfo contains the info of ports and other required information
-// needed to communicate with the beacon node it represents.
-type BeaconNodeInfo struct {
-	processID   int
-	datadir     string
-	rpcPort     uint64
-	monitorPort uint64
-	grpcPort    uint64
-	multiAddr   string
-}
 
 type end2EndConfig struct {
 	beaconFlags    []string
@@ -44,10 +30,10 @@ type end2EndConfig struct {
 var beaconNodeLogFileName = "beacon-%d.log"
 
 // startBeaconNodes starts the requested amount of beacon nodes, passing in the deposit contract given.
-func startBeaconNodes(t *testing.T, config *end2EndConfig) []*BeaconNodeInfo {
+func startBeaconNodes(t *testing.T, config *end2EndConfig) []*ev.BeaconNodeInfo {
 	numNodes := config.numBeaconNodes
 
-	nodeInfo := []*BeaconNodeInfo{}
+	nodeInfo := []*ev.BeaconNodeInfo{}
 	for i := uint64(0); i < numNodes; i++ {
 		newNode := startNewBeaconNode(t, config, nodeInfo)
 		nodeInfo = append(nodeInfo, newNode)
@@ -56,7 +42,7 @@ func startBeaconNodes(t *testing.T, config *end2EndConfig) []*BeaconNodeInfo {
 	return nodeInfo
 }
 
-func startNewBeaconNode(t *testing.T, config *end2EndConfig, beaconNodes []*BeaconNodeInfo) *BeaconNodeInfo {
+func startNewBeaconNode(t *testing.T, config *end2EndConfig, beaconNodes []*ev.BeaconNodeInfo) *ev.BeaconNodeInfo {
 	tmpPath := config.tmpPath
 	index := len(beaconNodes)
 	binaryPath, found := bazel.FindBinary("beacon-chain", "beacon-chain")
@@ -91,7 +77,7 @@ func startNewBeaconNode(t *testing.T, config *end2EndConfig, beaconNodes []*Beac
 	// After the first node is made, have all following nodes connect to all previously made nodes.
 	if index >= 1 {
 		for p := 0; p < index; p++ {
-			args = append(args, fmt.Sprintf("--peer=%s", beaconNodes[p].multiAddr))
+			args = append(args, fmt.Sprintf("--peer=%s", beaconNodes[p].MultiAddr))
 		}
 	}
 
@@ -112,13 +98,13 @@ func startNewBeaconNode(t *testing.T, config *end2EndConfig, beaconNodes []*Beac
 		t.Fatalf("could not get multiaddr for node %d: %v", index, err)
 	}
 
-	return &BeaconNodeInfo{
-		processID:   cmd.Process.Pid,
-		datadir:     fmt.Sprintf("%s/eth2-beacon-node-%d", tmpPath, index),
-		rpcPort:     4200 + uint64(index),
-		monitorPort: 8280 + uint64(index),
-		grpcPort:    3400 + uint64(index),
-		multiAddr:   multiAddr,
+	return &ev.BeaconNodeInfo{
+		ProcessID:   cmd.Process.Pid,
+		Datadir:     fmt.Sprintf("%s/eth2-beacon-node-%d", tmpPath, index),
+		RpcPort:     4200 + uint64(index),
+		MonitorPort: 8280 + uint64(index),
+		GrpcPort:    3400 + uint64(index),
+		MultiAddr:   multiAddr,
 	}
 }
 
@@ -140,34 +126,4 @@ func getMultiAddrFromLogFile(name string) (string, error) {
 		return "", fmt.Errorf("did not find peer text in %s", contents)
 	}
 	return contents[startIdx : startIdx+endIdx], nil
-}
-
-func waitForTextInFile(file *os.File, text string) error {
-	wait := 0
-	// Cap the wait in case there are issues starting.
-	maxWait := 36
-	for wait < maxWait {
-		time.Sleep(2 * time.Second)
-		// Rewind the file pointer to the start of the file so we can read it again.
-		_, err := file.Seek(0, io.SeekStart)
-		if err != nil {
-			return errors.Wrap(err, "could not rewind file to start")
-		}
-
-		scanner := bufio.NewScanner(file)
-		for scanner.Scan() {
-			if strings.Contains(scanner.Text(), text) {
-				return nil
-			}
-		}
-		if err := scanner.Err(); err != nil {
-			return err
-		}
-		wait += 2
-	}
-	contents, err := ioutil.ReadFile(file.Name())
-	if err != nil {
-		return err
-	}
-	return fmt.Errorf("could not find requested text \"%s\" in logs:\n%s", text, string(contents))
 }
