+++ endtoend/eth1.go
@@ -17,7 +17,6 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/ethereum/go-ethereum/rpc"
 	contracts "github.com/prysmaticlabs/prysm/contracts/deposit-contract"
@@ -119,38 +118,4 @@ func startEth1(t *testing.T, tmpPath string) (common.Address, string, int) {
 	}
 
 	return contractAddr, keystorePath, cmd.Process.Pid
-}
-
-func mineBlocks(web3 *ethclient.Client, keystore *keystore.Key, blocksToMake uint64) error {
-	nonce, err := web3.PendingNonceAt(context.Background(), keystore.Address)
-	if err != nil {
-		return err
-	}
-	chainID, err := web3.NetworkID(context.Background())
-	if err != nil {
-		return err
-	}
-	block, err := web3.BlockByNumber(context.Background(), nil)
-	if err != nil {
-		return err
-	}
-	finishBlock := block.NumberU64() + blocksToMake
-
-	for block.NumberU64() <= finishBlock {
-		spamTX := types.NewTransaction(nonce, keystore.Address, big.NewInt(0), 21000, big.NewInt(1e6), []byte{})
-		signed, err := types.SignTx(spamTX, types.NewEIP155Signer(chainID), keystore.PrivateKey)
-		if err != nil {
-			return err
-		}
-		if err := web3.SendTransaction(context.Background(), signed); err != nil {
-			return err
-		}
-		nonce++
-		time.Sleep(250 * time.Microsecond)
-		block, err = web3.BlockByNumber(context.Background(), nil)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
