+++ endtoend/evaluators/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     testonly = True,
     srcs = [
+        "consensus.go",
         "finality.go",
         "validator.go",
     ],
@@ -14,5 +15,6 @@ go_library(
         "@com_github_gogo_protobuf//types:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
+        "@org_golang_google_grpc//:go_default_library",
     ],
 )
